First working version

+++ src/com/cezar/homework/w04/Vehicle.java
@@ -3,6 +3,6 @@ package com.cezar.homework.w04;
 interface Vehicle {
 
     void start();
-    void drive(double kms);
+    void drive(float kms);
     void stop();
 }
